agents/fomc-research/pyproject.toml
[tool.poetry]
name = "fomc-research"
version = "0.1"
license = "Apache-2.0"
description = "FOMC Research Agent Demo using Google ADK"
authors = ["Mike Styer <styer@google.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
absl-py = "^2.2.1"
diff-match-patch = "^20241021"
google-adk = ">=0.0.2"
google-cloud-bigquery = "^3.30.0"
google-genai = "^1.5.0"
pdfplumber = "^0.11.5"
pydantic = "^2.10.6"
requests = "^2.32.3"
tabulate = "^0.9.0"
scikit-learn = "^1.6.1"
google-cloud-aiplatform = {extras = ["adk", "agent-engines"], version = "^1.88.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.25.3"
flake8-pyproject = "^1.2.3"
pylint = "^3.3.6"
pyink = "^24.10.1"

[tool.pytest.ini_options]
console_output_style = "progress"
addopts = "-vv -s"
testpaths = ["tests/"]
markers = [
    "unit"
]
log_level = "ERROR"
log_cli = false
log_auto_indent = true
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "[%(asctime)s] %(levelname)s (%(funcName)s) \t [%(pathname)s:%(lineno)d] %(message)s"
filterwarnings = [
    "ignore::UserWarning",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"